// Frontend/community.marketunderworld.com/src/lib/types.ts
export interface UserIdentity {
  ip: string;
  country: string;
  countryCode: string;
  region: string;
  regionId: string;
  city: string;
  language: string;
  timezone: string;
  detectedAt: string;
}